Develop

+++ server/managr/salesforce/background/__init__.py
@@ -163,8 +163,12 @@ def emit_add_products_to_sf(workflow_id, *args):
     return _process_add_products_to_sf(workflow_id, *args)
 
 
-def emit_generate_form_template(user_id):
-    return _generate_form_template(user_id)
+def emit_generate_form_template(user_id, delete_forms=False):
+    return _generate_form_template(user_id, delete_forms)
+
+
+def emit_generate_team_form_templates(user_id):
+    return _generate_team_form_templates(user_id)
 
 
 def emit_update_current_db_values(user_id, resource_type, integration_id, verbose_name):
@@ -226,16 +230,16 @@ def _process_gen_next_object_field_sync(user_id, operations_list, for_dev):
 
 @background()
 @log_all_exceptions
-def _generate_form_template(user_id):
+def _generate_form_template(user_id, delete_forms):
     user = User.objects.get(id=user_id)
     org = user.organization
     # delete all existing forms
-
-    org.custom_slack_forms.all().delete()
+    if delete_forms:
+        org.custom_slack_forms.all().delete()
     for form in slack_consts.INITIAL_FORMS:
         resource, form_type = form.split(".")
         f = OrgCustomSlackForm.objects.create(
-            form_type=form_type, resource=resource, organization=org
+            form_type=form_type, resource=resource, organization=org, team=user.team
         )
         public_fields = SObjectField.objects.filter(
             is_public=True,
@@ -249,6 +253,30 @@ def _generate_form_template(user_id):
             elif i == 1 and note is not None:
                 f.fields.add(note, through_defaults={"order": i})
         f.save()
+
+
+@background()
+@log_all_exceptions
+def _generate_team_form_templates(user_id):
+    from managr.organization.models import Team
+
+    user = User.objects.get(id=user_id)
+    org = user.organization
+    team_ref = (
+        Team.objects.filter(organization=user.organization).order_by("datetime_created").first()
+    )
+    forms = team_ref.team_forms.all()
+    for form in forms:
+        f = OrgCustomSlackForm.objects.create(
+            form_type=form.form_type,
+            resource=form.resource,
+            organization=org,
+            team=user.team,
+            config=form.config,
+            stage=form.stage,
+        )
+        if len(f.config):
+            f.recreate_form()
 
 
 @background(schedule=0, queue=sf_consts.SALESFORCE_RESOURCE_SYNC_QUEUE)
